// src/app/interfaces/formacion.ts
export interface Carreras {
    institucion: string;
    nivel:       string;
    titulo:      string;
    inicio:      number;
    fin:         number;
    estado:      boolean;
    cursando:    boolean;
  }
  
  export interface Cursos {
    titulo:      string;
    descripcion: string;
    lugar:       string;
    anio:        number;
  }
  